requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.59.0
debugpy==1.8.14
decorator==5.2.1
docker==7.1.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
Flask==3.1.1
fonttools==4.58.5
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazypredict==0.2.16
lightgbm==4.6.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mlflow==3.1.1
mlflow-skinny==3.1.1
narwhals==1.48.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.1
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
packaging==25.0
pandas==2.3.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
plotly-express==0.4.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.2
statsmodels==0.14.5
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xgboost==3.0.2
zipp==3.23.0
